Create a menu with interactive buttons

+++ PlatformerGame/src/entities/Player.java
@@ -39,7 +39,7 @@ public class Player extends Entity {
     public Player(float x, float y, int width, int height) {
         super(x, y, width, height);
         loadPlayerAnimations();
-        initHitbox(x, y, 20 * Game.SCALE, 28 * Game.SCALE);
+        initHitbox(x, y, (int) (20 * Game.SCALE), (int) (28 * Game.SCALE));
     }
 
     /**
